Wrapping up cleaning up NotificationCenter
